requirements.txt
einops==0.4.1
fastmri==0.2.0
fire==0.4.0
h5py==3.7.0
joblib==1.3.2
matplotlib==3.5.0
memory_profiler==0.61.0
numpy==1.25.2
numpy==1.23.5
opencv_python_headless==4.6.0.66
pandas==1.5.0
Pillow==9.4.0
Pillow==9.2.0
Pillow==10.2.0
pyarrow==15.0.0
pytorch_lightning==1.7.6
PyYAML==5.4.1
PyYAML==6.0.1
sacred==0.8.5
scikit_learn==1.0.2
timm==0.6.7
torch==2.0.1
torchmetrics==0.9.3
torchvision==0.13.0+cu102
torchvision==0.15.2
tqdm==4.65.0
transformers==4.22.1
